requirements.txt
#Django==1.7.1
djangorestframework==2.3.14
django-json-field==0.5.5
flake8==2.3.0
Cheetah==2.4.4
PyYAML==3.10
argparse==1.2.1
colorama==0.2.5
configobj==4.7.2
html5lib==0.999
jsonpatch==1.3
jsonpointer==1.0
mccabe==0.2.1
oauth==1.0.1
pep8==1.5.7
prettytable==0.7.2
psycopg2==2.5.2
pyflakes==0.8.1
pyserial==2.6
requests==2.4.3
six==1.5.2
ssh-import-id==3.21
urllib3==1.7.1
wsgiref==0.1.2
zope.interface==4.0.5
pytest==2.6.4
pytest-django==2.7.0